refactor: 이전에 VAC 패턴 적용했던 컴포넌트들 파일 분리
