+++ components/Common/PullToRefresh/PullToRefresh.tsx
@@ -1,11 +1,13 @@
 /* eslint-disable react/require-default-props */
-import { Box, Flex, Spinner } from '@chakra-ui/react';
+import { Spinner } from '@chakra-ui/react';
 import { useDrag } from '@use-gesture/react';
 import { useEffect, useRef, useState } from 'react';
 
 import { PullingIcon } from '@/public/icon';
 import getScrollParent from '@/utils/getScrollParent';
 import sleep from '@/utils/sleep';
+
+import PullToRefreshView from './PullToRefreshView';
 
 type PullStatus = 'pulling' | 'canRelease' | 'refreshing' | 'complete';
 
@@ -174,7 +176,7 @@ const PullToRefresh = ({
     }
   );
 
-  const PullToRefreshVAProps: PullToRefreshVAProps = {
+  const PullToRefreshVAProps = {
     status,
     pullingComponent,
     canReleaseComponent,
@@ -187,53 +189,6 @@ const PullToRefresh = ({
   };
 
   return <PullToRefreshView {...PullToRefreshVAProps} />;
-};
-
-interface PullToRefreshVAProps {
-  status: PullStatus;
-  pullingComponent: React.ReactNode;
-  canReleaseComponent: React.ReactNode;
-  refreshingComponent: React.ReactNode;
-  completeComponent: React.ReactNode;
-  elementRef: React.RefObject<HTMLDivElement>;
-  headRef: React.RefObject<HTMLDivElement>;
-  headHeight: number;
-  children: React.ReactNode;
-}
-
-const PullToRefreshView = ({
-  status,
-  elementRef,
-  headRef,
-  headHeight,
-  children,
-  ...restProps
-}: PullToRefreshVAProps) => {
-  const renderStatusComponent = {
-    pulling: restProps.pullingComponent,
-    canRelease: restProps.canReleaseComponent,
-    refreshing: restProps.refreshingComponent,
-    complete: restProps.completeComponent,
-  };
-  return (
-    <Box ref={elementRef}>
-      <Box ref={headRef} overflow="hidden" pos="relative">
-        <Flex
-          pos="absolute"
-          bottom={0}
-          left={0}
-          w="100%"
-          mb="20px"
-          h={headHeight}
-          align="flex-end"
-          justify="center"
-        >
-          {renderStatusComponent[status]}
-        </Flex>
-      </Box>
-      <Box>{children}</Box>
-    </Box>
-  );
 };
 
 export default PullToRefresh;
