+++ components/Main/LinkedFilter/LinkedFilter.tsx
@@ -1,0 +1,16 @@
+import useControlFilter from '@/hooks/useControlFilter';
+
+import LinkedFilterView from './LinkedFilterView';
+
+const LinkedFilter = () => {
+  const { handleClickFilter, currentValue } = useControlFilter();
+
+  const LinkedFilterVAProps = {
+    handleClickFilter,
+    currentValue,
+  };
+
+  return <LinkedFilterView {...LinkedFilterVAProps} />;
+};
+
+export default LinkedFilter;
