+++ components/Main/ProgressingStamps/ProgressingStamps.tsx
@@ -1,18 +1,15 @@
 import 'react-spring-bottom-sheet/dist/style.css';
 import 'swiper/css';
 
-import { Box, Skeleton, Text, VStack } from '@chakra-ui/react';
 import axios from 'axios';
 import { useEffect, useState } from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
-import { EffectCoverflow } from 'swiper';
-import { Swiper, SwiperSlide } from 'swiper/react';
 
 import { filterAtom } from '@/store/filter';
 import userInfoAtom from '@/store/userInfo';
 
-import PullToRefresh from '../Common/PullToRefresh';
-import Card from './Card';
+import ProgressingStampsSkeleton from './ProgressingStampsSkeleton';
+import ProgressingStampsView from './ProgressingStampsView';
 
 interface StampData {
   nickname: string;
@@ -190,121 +187,5 @@ const ProgressingStamps = () => {
     <ProgressingStampsView {...ProgressingStampsVAProps} />
   );
 };
-
-interface ProgressingStampsVAProps {
-  handleRefresh: () => Promise<any>;
-  cards: StampData[];
-  filter: string;
-}
-
-const ProgressingStampsView = ({
-  handleRefresh,
-  cards,
-  filter,
-}: ProgressingStampsVAProps) => (
-  <PullToRefresh onRefresh={handleRefresh}>
-    {filter !== '전체' ? (
-      <VStack w="100%" spacing="20px">
-        {cards
-          .find(({ nickname }) => nickname === filter)
-          ?.stamps.map(
-            ({ id, title, currentStamp, totalStamp, requestCount, reward }) => (
-              <Card
-                key={id}
-                title={title}
-                currentStamp={currentStamp}
-                totalStamp={totalStamp}
-                requestCount={requestCount}
-                reward={reward}
-              />
-            )
-          )}
-      </VStack>
-    ) : (
-      totalStampData.map(({ nickname, stamps }) => (
-        <Box key={nickname}>
-          <Text layerStyle="head20B" mb="13px">
-            {nickname}
-            <Text as="span" layerStyle="body18R">
-              님과 함께해요
-            </Text>
-          </Text>
-          <Swiper
-            effect="coverflow"
-            grabCursor
-            modules={[EffectCoverflow]}
-            slidesPerView={1.05}
-            centeredSlides
-            spaceBetween={0}
-            coverflowEffect={{
-              rotate: 10, // 회전각도
-              stretch: 0,
-              depth: 100, // 깊이감도
-              modifier: 2, //
-              slideShadows: false,
-            }}
-            style={{ marginBottom: '38px' }}
-          >
-            {stamps.map(
-              ({
-                id,
-                title,
-                currentStamp,
-                totalStamp,
-                requestCount,
-                reward,
-              }) => (
-                <SwiperSlide key={id}>
-                  <Card
-                    title={title}
-                    currentStamp={currentStamp}
-                    totalStamp={totalStamp}
-                    requestCount={requestCount}
-                    reward={reward}
-                  />
-                </SwiperSlide>
-              )
-            )}
-          </Swiper>
-        </Box>
-      ))
-    )}
-  </PullToRefresh>
-);
-
-interface ProgressingStampsSkeletonProps {
-  filter: string;
-}
-
-const ProgressingStampsSkeleton = ({
-  filter,
-}: ProgressingStampsSkeletonProps) =>
-  filter === '전체' ? (
-    <VStack w="100%" spacing="32px">
-      <VStack w="100%" spacing="13px">
-        <Skeleton
-          w="50%"
-          borderRadius="8px"
-          height="30px"
-          alignSelf="flex-start"
-        />
-        <Skeleton w="100%" borderRadius="8px" height="400px" />
-      </VStack>
-      <VStack w="100%" spacing="13px">
-        <Skeleton
-          w="50%"
-          borderRadius="8px"
-          height="30px"
-          alignSelf="flex-start"
-        />
-        <Skeleton w="100%" borderRadius="8px" height="400px" />
-      </VStack>
-    </VStack>
-  ) : (
-    <VStack w="100%" spacing="20px">
-      <Skeleton w="100%" borderRadius="8px" height="400px" />
-      <Skeleton w="100%" borderRadius="8px" height="400px" />
-    </VStack>
-  );
 
 export default ProgressingStamps;
